+++ src/components/Login&Signup/Signup.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { GoogleLogin } from '.';
-import userServices from '../utils/userServices';
+import userServices from '../../utils/userServices';
+import GoogleLogin from './GoogleLogin';
 import Swal from 'sweetalert2';
-import '../css/Login.css';
+import '../../css/Signup.css';
 
-class Login extends Component {
+class Signup extends Component {
   static propTypes = {
-    handleLogin: PropTypes.func.isRequired,
     history: PropTypes.shape({
       push: PropTypes.func,
     }).isRequired,
@@ -18,7 +17,9 @@ class Login extends Component {
     super(props);
     this.state = {
       email: '',
+      fullName: '',
       pw: '',
+      pwConfirm: '',
     };
   }
 
@@ -28,72 +29,95 @@ class Login extends Component {
     });
   };
 
+  /**
+   * submit user information to signup
+   * if successful redirect to homepage
+   */
   handleSubmit = e => {
-    const { handleLogin, history } = this.props;
+    const { handleSignup, history } = this.props;
     e.preventDefault();
     userServices
-      .login(this.state)
+      .signup(this.state)
       .then(() => {
-        handleLogin();
+        handleSignup();
         Swal.fire({
           position: 'top-end',
           type: 'success',
-          text: 'Successful Login',
+          text: 'Successful signup',
           showConfirmButton: false,
           timer: 1000,
         });
         history.push('/');
       })
-      .catch(err => 
+      .catch(err =>
         Swal.fire({
           position: 'top-end',
           type: 'error',
           text: err,
           showConfirmButton: false,
           timer: 1000,
-        })
+        }),
       );
   };
 
+  isFormInvalid() {
+    const { email, fullName, pw, pwConfirm } = this.state;
+    return !(fullName && email && pw === pwConfirm);
+  }
+
   render() {
-    const { email, pw, message } = this.state;
+    const { email, fullName, pw, pwConfirm } = this.state;
     return (
-      <div className="login-page">
-        <div className="login">
+      <div className="signup-page">
+        <div className="signup">
           <div className="form-container">
             <form onSubmit={this.handleSubmit}>
               <img
                 className="popcorn-logo"
                 src="/images/popcornKernal.png"
-                alt="YaMovie Popcorn"
+                alt="Broken"
               />
-              <header>Log In To YaMovie!</header>
+              <header>Signup with YaMovie!</header>
               <div className="input-container">
                 <input
-                  className="login-input"
+                  className="signup-input"
                   type="email"
                   placeholder="Email"
                   value={email}
                   onChange={e => this.handleChange('email', e)}
                 />
                 <input
-                  className="login-input"
+                  className="signup-input"
+                  type="name"
+                  placeholder="First and Last Name"
+                  value={fullName}
+                  onChange={e => this.handleChange('fullName', e)}
+                />
+                <input
+                  className="signup-input"
                   type="password"
                   placeholder="Password"
                   value={pw}
                   onChange={e => this.handleChange('pw', e)}
                 />
+                <input
+                  className="signup-input"
+                  type="password"
+                  placeholder="Password Confirmation"
+                  value={pwConfirm}
+                  onChange={e => this.handleChange('pwConfirm', e)}
+                />
               </div>
               <div className="button-container">
                 <button
-                  className="login-submit"
+                  className="signup-submit"
                   type="button"
                   onClick={this.handleSubmit}
                 >
                   {' '}
-                  Log In{' '}
+                  Signup{' '}
                 </button>
-                <Link className="login-cancel" to="/">
+                <Link className="signup-cancel" to="/">
                   Cancel
                 </Link>
               </div>
@@ -106,4 +130,8 @@ class Login extends Component {
   }
 }
 
-export default Login;
+Signup.propTypes = {
+  handleSignup: PropTypes.func.isRequired,
+};
+
+export default Signup;
