+++ src/components/index.js
@@ -1,16 +1,16 @@
-export { default as AccountNav } from './AccountNav';
-export { default as ChatWindow } from './ChatWindow';
+export { default as AccountNav } from './Account/AccountNav';
+export { default as ChatWindow } from './MovieDisplays/ChatWindow';
 export { default as FeedbackToast } from './FeedbackToast';
-export { default as GenreList } from './GenreList';
-export { default as GoogleLogin } from './GoogleLogin';
-export { default as Login } from './Login';
-export { default as MovieCard } from './MovieCard';
-export { default as MovieFeed } from './MovieFeed';
-export { default as MovieFeedItem } from './MovieFeedItem';
-export { default as MovieList } from './MovieList';
+export { default as GenreList } from './MovieDisplays/GenreList';
+export { default as GoogleLogin } from './Login&Signup/GoogleLogin';
+export { default as Login } from './Login&Signup/Login';
+export { default as MovieCard } from './MovieDisplays/MovieCard';
+export { default as MovieFeed } from './MovieDisplays/MovieFeed';
+export { default as MovieFeedItem } from './MovieDisplays/MovieFeedItem';
+export { default as MovieList } from './MovieDisplays/MovieList';
 export { default as Navbar } from './Navbar';
-export { default as OAuth } from './OAuth';
-export { default as SearchBar } from './SearchBar';
-export { default as Signup } from './Signup';
-export { default as TrailerModal } from './TrailerModal';
-export { default as Watchlist } from './Watchlist';
+export { default as OAuth } from './Login&Signup/OAuth';
+export { default as SearchBar } from './MovieDisplays/SearchBar';
+export { default as Signup } from './Login&Signup/Signup';
+export { default as TrailerModal } from './MovieDisplays/TrailerModal';
+export { default as Watchlist } from './Account/Watchlist';
