+++ src/pages/UserDashboardPage.js
@@ -3,8 +3,8 @@ import React, { Component } from 'react';
 import '../css/main.css';
 import '../css/UserDashboardPage.css';
 import { Route } from 'react-router-dom';
-import AccountNav from '../components/AccountNav';
-import UserPreferences from '../components/UserPreferences';
+import AccountNav from '../components/Account/AccountNav';
+import UserPreferences from '../components//Account/UserPreferences';
 
 class userDashboardPage extends Component {
   render() {
@@ -13,15 +13,18 @@ class userDashboardPage extends Component {
       <div>
         <div className="account-wrapper">
           <AccountNav />
-          
-          <Route exact path={`${match.url}`} render={({ props }) =>  <UserPreferences {...props} user={user} />} />
+
+          <Route
+            exact
+            path={`${match.url}`}
+            render={({ props }) => <UserPreferences {...props} user={user} />}
+          />
           {/* 
           TO DO: implement watchlist
            */}
           {/* <Route path={`${match.url}/:watchlist`} component={Watchlist} /> */}
         </div>
       </div>
-
     );
   }
 }
